requirements.txt
streamlit
uuid
Pillow
qrcode
